% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/055-requested-api-completion-0-7.R
\name{algorithm_facet_effects}
\alias{algorithm_facet_effects}
\title{Extract algorithm facet effects}
\usage{
algorithm_facet_effects(object, channel = c("response", "process"))
}
\arguments{
\item{object}{A fitted eyeprocess model or audit object.}

\item{channel}{Measurement channel to inspect.}
}
\description{
Extract algorithm facet effects
}
