% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/071-api-lifecycle-0-9.R
\name{api_lifecycle_diff}
\alias{api_lifecycle_diff}
\title{Compare two API lifecycle registries}
\usage{
api_lifecycle_diff(old, new)
}
\arguments{
\item{old}{Old registry.}

\item{new}{New registry.}
}
\description{
Compare two API lifecycle registries
}
