% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/070-governed-pipelines-0-9.R
\name{audit_eye_pipeline}
\alias{audit_eye_pipeline}
\title{Audit a pipeline definition or completed run}
\usage{
audit_eye_pipeline(x)
}
\arguments{
\item{x}{Pipeline or pipeline run.}
}
\description{
Audit a pipeline definition or completed run
}
