% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/095-multimodal-measurement-contract-0-10.R
\name{audit_multimodal_measurement}
\alias{audit_multimodal_measurement}
\title{Audit a multimodal measurement object}
\usage{
audit_multimodal_measurement(x)
}
\arguments{
\item{x}{An `eye_multimodal_measurement`.}
}
\value{
An `eye_multimodal_audit`.
}
\description{
Audit a multimodal measurement object
}
