% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/058-process-preflight-governance-0-8.R
\name{audit_multivariate_process_quality}
\alias{audit_multivariate_process_quality}
\title{Alias emphasizing data-quality interpretation of process anomaly auditing}
\usage{
audit_multivariate_process_quality(...)
}
\arguments{
\item{...}{Additional arguments passed to the underlying method or helper.}
}
\description{
Alias emphasizing data-quality interpretation of process anomaly auditing
}
