% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/021-validation-program.R
\name{benchmark_eyeprocess}
\alias{benchmark_eyeprocess}
\title{Benchmark an eyeprocess operation}
\usage{
benchmark_eyeprocess(expr, iterations = 5L, label = "operation")
}
\arguments{
\item{expr}{Function with no arguments to benchmark.}

\item{iterations}{Number of iterations.}

\item{label}{Benchmark label.}
}
\value{
An `eye_benchmark` data frame.
}
\description{
Benchmark an eyeprocess operation
}
