% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/060-process-window-representations-0-8.R
\name{bind_process_windows}
\alias{bind_process_windows}
\title{Bind compatible process-window objects}
\usage{
bind_process_windows(...)
}
\arguments{
\item{...}{Additional arguments passed to the underlying method or helper.}
}
\description{
Bind compatible process-window objects
}
