% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/067-advanced-sensitivity-diagnostics-0-8.R
\name{biometric_imputation_sensitivity}
\alias{biometric_imputation_sensitivity}
\title{Run biometric-feature imputation as a sensitivity analysis}
\usage{
biometric_imputation_sensitivity(
  data,
  variables,
  methods = c("mice", "missForest"),
  m = 3L,
  maxit = 3L,
  seed = 521
)
}
\arguments{
\item{data}{Data containing biometric/process features.}

\item{variables}{Variables to impute.}

\item{methods}{Any of `mice` and `missForest`.}

\item{m}{Number of MICE imputations.}

\item{maxit}{Iteration count.}

\item{seed}{Seed.}
}
\value{
An `eye_biometric_imputation_sensitivity` object. Complete-case
  analysis is not replaced automatically.
}
\description{
Run biometric-feature imputation as a sensitivity analysis
}
