% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/060-process-window-representations-0-8.R
\name{compare_aoi_trajectories}
\alias{compare_aoi_trajectories}
\title{Compare AOI trajectory feature objects}
\usage{
compare_aoi_trajectories(...)
}
\arguments{
\item{...}{Additional arguments passed to the underlying method or helper.}
}
\description{
Compare AOI trajectory feature objects
}
