% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/061-pupil-advanced-representations-0-8.R
\name{compare_signal_filters}
\alias{compare_signal_filters}
\title{Compare multiple signal filters}
\usage{
compare_signal_filters(
  signal,
  widths = c(5L, 9L, 15L),
  methods = c("runmed", "robfilter")
)
}
\arguments{
\item{signal}{Numeric signal.}

\item{widths}{Widths to compare.}

\item{methods}{Methods to compare.}
}
\value{
A tabular R object containing multiple signal filters; rows represent analysis units and columns contain the returned quantities.
}
\description{
Compare multiple signal filters
}
