% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/073-decision-manifests-0-9.R
\name{decision_manifest_hash}
\alias{decision_manifest_hash}
\title{Stable hash of decision content}
\usage{
decision_manifest_hash(x)
}
\arguments{
\item{x}{Manifest or decision object.}
}
\description{
Stable hash of decision content
}
