% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/050-process-irt-models-0-7.R
\name{detect_irt_changepoints}
\alias{detect_irt_changepoints}
\title{Detect IRT/process change points using an SIC-inspired multichannel score}
\usage{
detect_irt_changepoints(
  data,
  person = "participant_id",
  order = "item_order",
  response = "response",
  rt = "rt",
  gaze = NULL,
  min_segment = 5L,
  min_delta_sic = 2,
  max_changes = 2L
)
}
\arguments{
\item{data}{Input data frame or compatible tabular object.}

\item{person}{Person or participant identifier column.}

\item{order}{Within-sequence ordering variable.}

\item{response}{Response variable or response-column name.}

\item{rt}{Response-time variable or column name.}

\item{gaze}{Gaze/process variable or column name.}

\item{min_segment}{Minimum segment length.}

\item{min_delta_sic}{Minimum information-criterion improvement.}

\item{max_changes}{Maximum number of change points.}
}
\description{
This implementation is a transparent package reference inspired by the 2026
SIC-CPA literature. It combines Bernoulli response likelihood with normal
log-RT and standardized gaze likelihoods. It is not a line-for-line
reproduction of the article's estimator.
}
