% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/055-requested-api-completion-0-7.R
\name{device_facet_effects}
\alias{device_facet_effects}
\title{Extract device facet effects}
\usage{
device_facet_effects(object, channel = c("response", "process"))
}
\arguments{
\item{object}{A fitted eyeprocess model or audit object.}

\item{channel}{Measurement channel to inspect.}
}
\description{
Extract device facet effects
}
