% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/020-interoperability-storage.R
\name{external_eye_adapters}
\alias{external_eye_adapters}
\alias{as_eyeprocess_eyetools}
\alias{as_eyeprocess_eyetrackingr}
\alias{as_eyeprocess_gazer}
\alias{as_eyeprocess_eyeris}
\alias{as_eyeprocess_pupillometryr}
\title{Convert common external eye-tracking objects}
\usage{
as_eyeprocess_eyetools(x, mapping = NULL, ...)

as_eyeprocess_eyetrackingr(x, mapping = NULL, ...)

as_eyeprocess_gazer(x, mapping = NULL, ...)

as_eyeprocess_eyeris(x, mapping = NULL, ...)

as_eyeprocess_pupillometryr(x, mapping = NULL, ...)
}
\arguments{
\item{x}{External object coercible to a data frame.}

\item{mapping}{Optional `eye_mapping`.}

\item{...}{Passed to `read_eye_generic()`.}
}
\value{
An `eye_dataset`.
}
\description{
Convert common external eye-tracking objects
}
