% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/071-api-lifecycle-0-9.R
\name{eye_api_recommendation}
\alias{eye_api_recommendation}
\title{Lifecycle recommendation for API review}
\usage{
eye_api_recommendation(audit)
}
\arguments{
\item{audit}{`eye_api_audit` object.}
}
\description{
Lifecycle recommendation for API review
}
