% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/070-governed-pipelines-0-9.R
\name{eye_pipeline_dot}
\alias{eye_pipeline_dot}
\title{Render pipeline dependencies as Graphviz DOT}
\usage{
eye_pipeline_dot(x)
}
\arguments{
\item{x}{Pipeline.}
}
\description{
Render pipeline dependencies as Graphviz DOT
}
