% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/070-governed-pipelines-0-9.R
\name{eye_pipeline_manifest}
\alias{eye_pipeline_manifest}
\title{Machine-readable pipeline manifest}
\usage{
eye_pipeline_manifest(x)
}
\arguments{
\item{x}{Pipeline.}
}
\description{
Machine-readable pipeline manifest
}
