% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/090-irt-multidimensional-cdm-0-9.R
\name{eyeprocess_cdm_attribute_profiles}
\alias{eyeprocess_cdm_attribute_profiles}
\title{Enumerate latent attribute profiles}
\usage{
eyeprocess_cdm_attribute_profiles(
  n_attributes,
  attribute_names = paste0("A", seq_len(n_attributes))
)
}
\arguments{
\item{n_attributes}{Number of cognitive-diagnosis attributes.}

\item{attribute_names}{Optional names for the cognitive-diagnosis attributes.}
}
\description{
Enumerate latent attribute profiles
}
