% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/090-irt-multidimensional-cdm-0-9.R
\name{eyeprocess_cdm_dina_ideal_response}
\alias{eyeprocess_cdm_dina_ideal_response}
\title{Compute deterministic DINA ideal responses from a Q-matrix}
\usage{
eyeprocess_cdm_dina_ideal_response(Q, profiles)
}
\arguments{
\item{Q}{Binary item-by-attribute Q-matrix.}

\item{profiles}{Attribute mastery profiles, with rows representing profiles.}
}
\description{
Compute deterministic DINA ideal responses from a Q-matrix
}
