% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/083-irt-foundations-information-0-9.R
\name{eyeprocess_irt_3pl_probability}
\alias{eyeprocess_irt_3pl_probability}
\title{3PL item-response probability}
\usage{
eyeprocess_irt_3pl_probability(theta, a = 1, b = 0, c = 0.2, D = 1)
}
\arguments{
\item{theta}{Latent-trait value or vector of latent-trait values.}

\item{a}{Item discrimination parameter or vector.}

\item{b}{Item difficulty or location parameter or vector.}

\item{c}{Lower-asymptote parameter or vector.}

\item{D}{Logistic scaling constant.}
}
\description{
3PL item-response probability
}
