% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/085-irt-scoring-adaptive-0-9.R
\name{eyeprocess_irt_adaptive_trace}
\alias{eyeprocess_irt_adaptive_trace}
\title{Create an auditable adaptive-testing trace}
\usage{
eyeprocess_irt_adaptive_trace(
  item_id,
  theta_before,
  theta_after,
  se_after,
  information,
  response = NA_real_
)
}
\arguments{
\item{item_id}{Item identifier or vector of item identifiers.}

\item{theta_before}{Latent-trait estimate before item administration.}

\item{theta_after}{Latent-trait estimate after item administration.}

\item{se_after}{Conditional standard error after item administration.}

\item{information}{Item or test information value or vector.}

\item{response}{Observed item response or response variable.}
}
\value{
An object of class "eye_irt_adaptive_trace", "data.frame", stored as a data frame, containing an auditable adaptive-testing trace and associated metadata needed to interpret the result.
}
\description{
Create an auditable adaptive-testing trace
}
