% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/088-irt-engine-adapters-0-9.R
\name{eyeprocess_irt_engine_registry}
\alias{eyeprocess_irt_engine_registry}
\title{Registry of specialized external IRT engines}
\usage{
eyeprocess_irt_engine_registry()
}
\value{
An object of class "eye_irt_engine_registry", "data.frame", stored as a data frame, containing registry of specialized external IRT engines and associated metadata needed to interpret the result.
}
\description{
Registry of specialized external IRT engines
}
