% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/083-irt-foundations-information-0-9.R
\name{eyeprocess_irt_gpcm_probability}
\alias{eyeprocess_irt_gpcm_probability}
\title{Generalized partial-credit category probabilities}
\usage{
eyeprocess_irt_gpcm_probability(theta, a = 1, steps, D = 1)
}
\arguments{
\item{theta}{Latent-trait value or vector of latent-trait values.}

\item{a}{Item discrimination parameter or vector.}

\item{steps}{Step parameters for a generalized partial-credit model.}

\item{D}{Logistic scaling constant.}
}
\value{
A logical value or vector indicating generalized partial-credit category probabilities.
}
\description{
Generalized partial-credit category probabilities
}
