% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/085-irt-scoring-adaptive-0-9.R
\name{eyeprocess_irt_information_gain}
\alias{eyeprocess_irt_information_gain}
\title{Information gain between two conditional standard errors}
\usage{
eyeprocess_irt_information_gain(se_before, se_after)
}
\arguments{
\item{se_before}{Conditional standard error before an item is administered.}

\item{se_after}{Conditional standard error after item administration.}
}
\description{
Information gain between two conditional standard errors
}
