% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/086-irt-linking-invariance-0-9.R
\name{eyeprocess_irt_mean_sigma_link}
\alias{eyeprocess_irt_mean_sigma_link}
\title{Mean-sigma IRT linking coefficients}
\usage{
eyeprocess_irt_mean_sigma_link(reference, focal, anchors = NULL)
}
\arguments{
\item{reference}{Reference-form or reference-group item parameters.}

\item{focal}{Focal-form or focal-group item parameters.}

\item{anchors}{Anchor-item identifiers.}
}
\description{
Mean-sigma IRT linking coefficients
}
