% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/085-irt-scoring-adaptive-0-9.R
\name{eyeprocess_irt_mle_score}
\alias{eyeprocess_irt_mle_score}
\title{Bounded ML score for dichotomous IRT item parameters}
\usage{
eyeprocess_irt_mle_score(response, items, bounds = c(-6, 6), D = 1)
}
\arguments{
\item{response}{Observed item response or response variable.}

\item{items}{Item-parameter data frame or item collection.}

\item{bounds}{Numerical lower and upper optimization bounds.}

\item{D}{Logistic scaling constant.}
}
\description{
Bounded ML score for dichotomous IRT item parameters
}
