% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/086-irt-linking-invariance-0-9.R
\name{eyeprocess_irt_process_dif_concordance}
\alias{eyeprocess_irt_process_dif_concordance}
\title{Compare psychometric DIF effect sizes with process-channel contrasts}
\usage{
eyeprocess_irt_process_dif_concordance(
  dif,
  process,
  item_id = "item_id",
  dif_effect = "effect",
  process_effect = "effect"
)
}
\arguments{
\item{dif}{Differential-item-functioning evidence or summary.}

\item{process}{Process-measure columns or process object.}

\item{item_id}{Item identifier or vector of item identifiers.}

\item{dif_effect}{DIF effect used for concordance.}

\item{process_effect}{Process-side item effect used for concordance.}
}
\description{
Compare psychometric DIF effect sizes with process-channel contrasts
}
