% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/083-irt-foundations-information-0-9.R
\name{eyeprocess_irt_sparse_design_audit}
\alias{eyeprocess_irt_sparse_design_audit}
\title{Audit sparse person-item response coverage}
\usage{
eyeprocess_irt_sparse_design_audit(
  data,
  person,
  item,
  response = NULL,
  min_person_items = 3L,
  min_item_persons = 10L
)
}
\arguments{
\item{data}{Input data frame, matrix, or compatible analysis object.}

\item{person}{Name of the person identifier column.}

\item{item}{Name of the item identifier column.}

\item{response}{Observed item response or response variable.}

\item{min_person_items}{Minimum number of observed items required per person.}

\item{min_item_persons}{Minimum number of observed persons required per item.}
}
\description{
Audit sparse person-item response coverage
}
