% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/085-irt-scoring-adaptive-0-9.R
\name{eyeprocess_irt_stopping_rule}
\alias{eyeprocess_irt_stopping_rule}
\title{Evaluate a simple adaptive stopping rule}
\usage{
eyeprocess_irt_stopping_rule(
  n_administered,
  se = NA_real_,
  min_items = 5L,
  max_items = 30L,
  target_se = 0.3
)
}
\arguments{
\item{n_administered}{Number of items already administered.}

\item{se}{Standard-error values.}

\item{min_items}{Minimum number of items required.}

\item{max_items}{Maximum permitted test length.}

\item{target_se}{Target conditional standard error for stopping.}
}
\description{
Evaluate a simple adaptive stopping rule
}
