% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/083-irt-foundations-information-0-9.R
\name{eyeprocess_irt_test_information}
\alias{eyeprocess_irt_test_information}
\title{Compute a test information curve from item parameters}
\usage{
eyeprocess_irt_test_information(theta, items, D = 1)
}
\arguments{
\item{theta}{Latent-trait value or vector of latent-trait values.}

\item{items}{Item-parameter data frame or item collection.}

\item{D}{Logistic scaling constant.}
}
\value{
An object of class "eye_irt_information_profile", "data.frame", stored as a data frame, containing a test information curve from item parameters and associated metadata needed to interpret the result.
}
\description{
Compute a test information curve from item parameters
}
