% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/090-irt-multidimensional-cdm-0-9.R
\name{eyeprocess_mirt_directional_information}
\alias{eyeprocess_mirt_directional_information}
\title{Directional multidimensional 2PL information}
\usage{
eyeprocess_mirt_directional_information(
  theta,
  discrimination,
  difficulty = 0,
  direction = NULL,
  D = 1
)
}
\arguments{
\item{theta}{Latent-trait value or vector of latent-trait values.}

\item{discrimination}{Discrimination vector or matrix.}

\item{difficulty}{Item difficulty or location parameter.}

\item{direction}{Direction vector used to project multidimensional information.}

\item{D}{Logistic scaling constant.}
}
\description{
Directional multidimensional 2PL information
}
