% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/090-irt-multidimensional-cdm-0-9.R
\name{eyeprocess_mirt_information_matrix}
\alias{eyeprocess_mirt_information_matrix}
\title{Multidimensional 2PL item information matrix}
\usage{
eyeprocess_mirt_information_matrix(
  theta,
  discrimination,
  difficulty = 0,
  D = 1
)
}
\arguments{
\item{theta}{Latent-trait value or vector of latent-trait values.}

\item{discrimination}{Discrimination vector or matrix.}

\item{difficulty}{Item difficulty or location parameter.}

\item{D}{Logistic scaling constant.}
}
\value{
A numeric value or vector containing multidimensional 2PL item information matrix.
}
\description{
Multidimensional 2PL item information matrix
}
