% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/087-irt-process-joint-0-9.R
\name{eyeprocess_process_item_profile}
\alias{eyeprocess_process_item_profile}
\title{Aggregate process channels by item}
\usage{
eyeprocess_process_item_profile(data, item, channels)
}
\arguments{
\item{data}{Input data frame, matrix, or compatible analysis object.}

\item{item}{Name of the item identifier column.}

\item{channels}{Names or definitions of measurement channels.}
}
\description{
Aggregate process channels by item
}
