% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/061-pupil-advanced-representations-0-8.R
\name{filter_pupil_signal}
\alias{filter_pupil_signal}
\title{Filter pupil signal robustly}
\usage{
filter_pupil_signal(...)
}
\arguments{
\item{...}{Additional arguments passed to the underlying method or helper.}
}
\description{
Filter pupil signal robustly
}
