% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/065-frontier-gates-0-8.R
\name{fit_crossclassified_process_irt_mhrm}
\alias{fit_crossclassified_process_irt_mhrm}
\title{Create a gated scalable cross-classified MH-RM process IRT interface}
\usage{
fit_crossclassified_process_irt_mhrm(data, engine = NULL, ...)
}
\arguments{
\item{data}{Data supplied to an optional external engine.}

\item{engine}{Optional estimator implementing the intended scalable MH-RM model.}

\item{...}{Passed to engine.}
}
\description{
Create a gated scalable cross-classified MH-RM process IRT interface
}
