% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/020-interoperability-storage.R
\name{fit_diffirt_adapter}
\alias{fit_diffirt_adapter}
\title{Fit a diffusion IRT adapter}
\usage{
fit_diffirt_adapter(x, model = c("D", "Q"), ...)
}
\arguments{
\item{x}{An `eye_dataset`.}

\item{model}{Diffusion IRT model, `"D"` or `"Q"`.}

\item{...}{Passed to `diffIRT::diffIRT()`.}
}
\value{
An `eyeprocess_model`.
}
\description{
Fit a diffusion IRT adapter
}
