% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/020-interoperability-storage.R
\name{fit_gdina_adapter}
\alias{fit_gdina_adapter}
\title{Fit a GDINA cognitive-diagnosis adapter}
\usage{
fit_gdina_adapter(x, q_matrix, model = "GDINA", ...)
}
\arguments{
\item{x}{An `eye_dataset`.}

\item{q_matrix}{Q-matrix with items in response-matrix order.}

\item{model}{GDINA model specification.}

\item{...}{Passed to `GDINA::GDINA()`.}
}
\value{
An `eyeprocess_model`.
}
\description{
Fit a GDINA cognitive-diagnosis adapter
}
