% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/051-advanced-process-irt-0-7.R
\name{fit_latent_class_process_irt}
\alias{fit_latent_class_process_irt}
\title{Latent process-class IRT reference model}
\usage{
fit_latent_class_process_irt(
  data,
  response = "response",
  process_features,
  person = "participant_id",
  item = "item_id",
  n_classes = 2L,
  seed = 1
)
}
\arguments{
\item{data}{Input data frame or compatible tabular object.}

\item{response}{Response variable or response-column name.}

\item{process_features}{Names of process-derived features.}

\item{person}{Person or participant identifier column.}

\item{item}{Item identifier, name, or item column.}

\item{n_classes}{Number of latent classes.}

\item{seed}{Random-number seed.}
}
\description{
Latent process-class IRT reference model
}
