% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/020-interoperability-storage.R
\name{fit_openmx_process_model}
\alias{fit_openmx_process_model}
\title{Fit a user-defined OpenMx process model}
\usage{
fit_openmx_process_model(x, model_builder, include_features = TRUE, ...)
}
\arguments{
\item{x}{An `eye_dataset`.}

\item{model_builder}{Function receiving model data and returning an OpenMx model.}

\item{include_features}{Whether to merge process features.}

\item{...}{Passed to `OpenMx::mxRun()`.}
}
\value{
An `eyeprocess_model`.
}
\description{
Fit a user-defined OpenMx process model
}
