% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/075-benchmark-stress-0-9.R
\name{inject_pupil_dropout}
\alias{inject_pupil_dropout}
\title{Inject pupil dropout}
\usage{
inject_pupil_dropout(data, pupil = "pupil", proportion, seed = 1L)
}
\arguments{
\item{data}{Data.}

\item{pupil}{Pupil column.}

\item{proportion}{Dropout proportion.}

\item{seed}{Seed.}
}
\description{
Inject pupil dropout
}
