% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/049-multimodal-irt-registry.R
\name{irt_sequence_channel}
\alias{irt_sequence_channel}
\title{Sequence/process-state channel}
\usage{
irt_sequence_channel(
  sequence = "sequence",
  family = c("ngram", "hmm", "embedding"),
  latent = "strategy",
  options = list()
)
}
\arguments{
\item{sequence}{Sequence input.}

\item{family}{Statistical family used by the channel or model.}

\item{latent}{Latent variable or latent-variable labels.}

\item{options}{Additional channel/model options.}
}
\description{
Sequence/process-state channel
}
