% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/073-decision-manifests-0-9.R
\name{lock_decision_manifest}
\alias{lock_decision_manifest}
\title{Lock a decision manifest by content hash}
\usage{
lock_decision_manifest(x, label = "analysis_decisions")
}
\arguments{
\item{x}{Manifest.}

\item{label}{Optional lock label.}
}
\description{
Lock a decision manifest by content hash
}
