% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/021-validation-program.R
\name{model_validation_summary}
\alias{model_validation_summary}
\title{Summarize model validation}
\usage{
model_validation_summary(x)
}
\arguments{
\item{x}{An `eye_model_validation` object.}
}
\value{
Scenario-by-parameter validation metrics.
}
\description{
Summarize model validation
}
