% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/062-context-process-structure-0-8.R
\name{multiblock_contributions}
\alias{multiblock_contributions}
\title{Extract multiblock block contributions/coordinates}
\usage{
multiblock_contributions(x)
}
\arguments{
\item{x}{Object to process, inspect, compare, or plot.}
}
\description{
Extract multiblock block contributions/coordinates
}
