% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/098-multimodal-fit-contract-0-10.R
\name{multimodal_ppc}
\alias{multimodal_ppc}
\title{Posterior predictive checks for multimodal development fits}
\usage{
multimodal_ppc(object, variables = NULL)
}
\arguments{
\item{object}{An `eye_multimodal_irt_fit`.}

\item{variables}{Optional generated-quantity variable names.}
}
\value{
An `eye_multimodal_ppc`.
}
\description{
Posterior predictive checks for multimodal development fits
}
