% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/077-negative-controls-leakage-0-9.R
\name{negative_control_concordance}
\alias{negative_control_concordance}
\title{Concordance of multiple negative-control families}
\usage{
negative_control_concordance(x, effect = "effect", tolerance = 0.05)
}
\arguments{
\item{x}{Negative-control result.}

\item{effect}{Effect column.}

\item{tolerance}{Absolute mean-effect tolerance.}
}
\description{
Concordance of multiple negative-control families
}
