% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/078-reproducibility-provenance-0-9.R
\name{object_hash}
\alias{object_hash}
\title{Hash an R object reproducibly within an R serialization version}
\usage{
object_hash(x)
}
\arguments{
\item{x}{R object.}
}
\description{
Hash an R object reproducibly within an R serialization version
}
