% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/050-process-irt-models-0-7.R
\name{option_process_information}
\alias{option_process_information}
\title{Quantify option-process information from a nominal gaze model}
\usage{
option_process_information(object)
}
\arguments{
\item{object}{A fitted eyeprocess model or audit object.}
}
\description{
Quantify option-process information from a nominal gaze model
}
