% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/070-governed-pipelines-0-9.R
\name{pipeline_step_status}
\alias{pipeline_step_status}
\title{Pipeline step status table}
\usage{
pipeline_step_status(x)
}
\arguments{
\item{x}{Pipeline run.}
}
\description{
Pipeline step status table
}
