% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/064-next-generation-plots-0-8.R
\name{plot.eye_aoi_trajectory}
\alias{plot.eye_aoi_trajectory}
\title{Plot aoi trajectory diagnostics}
\usage{
\method{plot}{eye_aoi_trajectory}(x, type = c("coefficients", "profiles"), ...)
}
\arguments{
\item{x}{Object to process, inspect, compare, or plot.}

\item{type}{Type of summary or visual representation to produce.}

\item{...}{Additional arguments passed to the underlying method or helper.}
}
\description{
Plot aoi trajectory diagnostics
}
