% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/053-process-irt-plots-0-7.R
\name{plot.eye_irt_ppc}
\alias{plot.eye_irt_ppc}
\title{Plot posterior predictive discrepancy tail probabilities}
\usage{
\method{plot}{eye_irt_ppc}(x, ...)
}
\arguments{
\item{x}{Object to print, plot, summarize, or audit.}

\item{...}{Additional arguments passed to the selected model, engine, or method.}
}
\description{
Plot posterior predictive discrepancy tail probabilities
}
