% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/053-process-irt-plots-0-7.R
\name{plot.eye_latent_space_irt}
\alias{plot.eye_latent_space_irt}
\title{Plot a latent-space IRT adapter fit}
\usage{
\method{plot}{eye_latent_space_irt}(x, ...)
}
\arguments{
\item{x}{Object to print, plot, summarize, or audit.}

\item{...}{Additional arguments passed to the selected model, engine, or method.}
}
\description{
Plot a latent-space IRT adapter fit
}
