% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/053-process-irt-plots-0-7.R
\name{plot.eye_manyfacet_process_irt}
\alias{plot.eye_manyfacet_process_irt}
\title{Plot many-facet process IRT effects}
\usage{
\method{plot}{eye_manyfacet_process_irt}(x, facet = NULL, ...)
}
\arguments{
\item{x}{Object to print, plot, summarize, or audit.}

\item{facet}{Facet to display or extract.}

\item{...}{Additional arguments passed to the selected model, engine, or method.}
}
\description{
Plot many-facet process IRT effects
}
