% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/053-process-irt-plots-0-7.R
\name{plot.eye_process_person_fit}
\alias{plot.eye_process_person_fit}
\title{Plot process person-fit discrepancies}
\usage{
\method{plot}{eye_process_person_fit}(x, top = 25L, ...)
}
\arguments{
\item{x}{Object to print, plot, summarize, or audit.}

\item{top}{Number of highest-ranked cases to display.}

\item{...}{Additional arguments passed to the selected model, engine, or method.}
}
\description{
Plot process person-fit discrepancies
}
