% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/064-next-generation-plots-0-8.R
\name{plot.eye_pupil_confound_model}
\alias{plot.eye_pupil_confound_model}
\title{Plot pupil confound model diagnostics}
\usage{
\method{plot}{eye_pupil_confound_model}(
  x,
  type = c("luminance", "trial_order", "raw_adjusted", "theta_luminance_surface"),
  ...
)
}
\arguments{
\item{x}{Object to process, inspect, compare, or plot.}

\item{type}{Type of summary or visual representation to produce.}

\item{...}{Additional arguments passed to the underlying method or helper.}
}
\description{
Plot pupil confound model diagnostics
}
