% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/064-next-generation-plots-0-8.R
\name{plot.eye_pupil_deconvolution}
\alias{plot.eye_pupil_deconvolution}
\title{Plot pupil deconvolution diagnostics}
\usage{
\method{plot}{eye_pupil_deconvolution}(x, type = c("observed_fitted", "effects", "residuals", "kernels"), ...)
}
\arguments{
\item{x}{Object to process, inspect, compare, or plot.}

\item{type}{Type of summary or visual representation to produce.}

\item{...}{Additional arguments passed to the underlying method or helper.}
}
\description{
Plot pupil deconvolution diagnostics
}
