% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/064-next-generation-plots-0-8.R
\name{plot.eye_pupil_frequency_stability}
\alias{plot.eye_pupil_frequency_stability}
\title{Plot pupil frequency stability diagnostics}
\usage{
\method{plot}{eye_pupil_frequency_stability}(x, feature = "pupil_frequency_contrast", ...)
}
\arguments{
\item{x}{Object to process, inspect, compare, or plot.}

\item{feature}{Process feature to evaluate or display.}

\item{...}{Additional arguments passed to the underlying method or helper.}
}
\description{
Plot pupil frequency stability diagnostics
}
