% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/055-requested-api-completion-0-7.R
\name{plot_person_item_space}
\alias{plot_person_item_space}
\title{Plot person/item latent-space coordinates}
\usage{
plot_person_item_space(object, dimensions = c(1L, 2L), labels = FALSE, ...)
}
\arguments{
\item{object}{Fitted `eye_latent_space_irt` object.}

\item{dimensions}{Two coordinate dimensions to display.}

\item{labels}{Whether to add entity labels.}

\item{...}{Graphical arguments.}
}
\description{
Plot person/item latent-space coordinates
}
