% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/064-next-generation-plots-0-8.R
\name{plot_pupil_activity_windows}
\alias{plot_pupil_activity_windows}
\title{Plot pupil activity features across windows/groups}
\usage{
plot_pupil_activity_windows(x, feature = "pupil_frequency_contrast", ...)
}
\arguments{
\item{x}{`eye_pupil_frequency_features` or `eye_pupil_frequency_stability` object.}

\item{feature}{Activity feature.}

\item{...}{Additional arguments passed to the underlying method or helper.}
}
\description{
Plot pupil activity features across windows/groups
}
