% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/064-next-generation-plots-0-8.R
\name{plot_pupil_band_power}
\alias{plot_pupil_band_power}
\title{Plot pupil low/high-band power summaries}
\usage{
plot_pupil_band_power(x, ...)
}
\arguments{
\item{x}{`eye_pupil_frequency_features` object.}

\item{...}{Additional arguments passed to the underlying method or helper.}
}
\description{
Plot pupil low/high-band power summaries
}
