% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/064-next-generation-plots-0-8.R
\name{plot_pupil_components}
\alias{plot_pupil_components}
\title{Plot tonic/phasic pupil components}
\usage{
plot_pupil_components(
  data,
  time = "time_ms",
  smoothed = "pupil_smoothed",
  tonic = "pupil_tonic",
  phasic = "pupil_phasic",
  ...
)
}
\arguments{
\item{data}{Data frame containing the required process variables.}

\item{time}{Time values or name of the time variable.}

\item{smoothed}{Name of the smoothed pupil-signal column.}

\item{tonic}{Name of the tonic pupil-component column.}

\item{phasic}{Name of the phasic pupil-component column.}

\item{...}{Additional arguments passed to the underlying method or helper.}
}
\description{
Plot tonic/phasic pupil components
}
