% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/051-advanced-process-irt-0-7.R
\name{predict_theta_at_time}
\alias{predict_theta_at_time}
\title{Predict a latent trait at arbitrary times}
\usage{
predict_theta_at_time(object, time)
}
\arguments{
\item{object}{A fitted eyeprocess model or audit object.}

\item{time}{Time values.}
}
\description{
Predict a latent trait at arbitrary times
}
