% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/058-process-preflight-governance-0-8.R
\name{preflight_failures}
\alias{preflight_failures}
\title{Extract pre-flight failures/review cases}
\usage{
preflight_failures(x)
}
\arguments{
\item{x}{Object to process, inspect, compare, or plot.}
}
\description{
Extract pre-flight failures/review cases
}
