% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/058-process-preflight-governance-0-8.R
\name{preflight_passed}
\alias{preflight_passed}
\title{Extract pre-flight passes}
\usage{
preflight_passed(x)
}
\arguments{
\item{x}{Object to process, inspect, compare, or plot.}
}
\value{
A logical value or vector indicating pre-flight passes.
}
\description{
Extract pre-flight passes
}
