% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/060-process-window-representations-0-8.R
\name{print.eye_process_window_spec}
\alias{print.eye_process_window_spec}
\title{Print a process window spec object}
\usage{
\method{print}{eye_process_window_spec}(x, ...)
}
\arguments{
\item{x}{Object to process, inspect, compare, or plot.}

\item{...}{Additional arguments passed to the underlying method or helper.}
}
\description{
Print a process window spec object
}
