% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/059-deployment-drift-0-8.R
\name{process_drift_alerts}
\alias{process_drift_alerts}
\title{Extract drift alerts}
\usage{
process_drift_alerts(x)
}
\arguments{
\item{x}{Object to process, inspect, compare, or plot.}
}
\value{
A logical value or vector indicating drift alerts.
}
\description{
Extract drift alerts
}
