% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/051-advanced-process-irt-0-7.R
\name{process_item_information}
\alias{process_item_information}
\title{2PL response item information}
\usage{
process_item_information(
  theta,
  a,
  b,
  process_information = 0,
  rt_information = 0,
  weights = c(response = 1, rt = 0, process = 0),
  expected_time = 0,
  burden_weight = 0
)
}
\arguments{
\item{theta}{Latent-trait values.}

\item{a}{Item discrimination parameter or parameters.}

\item{b}{Item difficulty/location parameter or parameters.}

\item{process_information}{Information supplied by the process channel.}

\item{rt_information}{Information supplied by response time.}

\item{weights}{Weights used to combine information components.}

\item{expected_time}{Expected response time or burden.}

\item{burden_weight}{Penalty applied to expected burden.}
}
\value{
An object of class "eye_process_item_information", stored as a named list, with components "theta", "response_information", "utility". It contains 2PL response item information and associated metadata or diagnostics needed to interpret the result.
}
\description{
2PL response item information
}
