% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/074-process-registry-reliability-0-9.R
\name{process_measure_lineage}
\alias{process_measure_lineage}
\title{Process-measure lineage table}
\usage{
process_measure_lineage(
  measure,
  inputs,
  transformations = character(),
  output_level = NA_character_
)
}
\arguments{
\item{measure}{Measure name.}

\item{inputs}{Required input variable names.}

\item{transformations}{Ordered transformation labels.}

\item{output_level}{Output aggregation level.}
}
\description{
Process-measure lineage table
}
