% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/074-process-registry-reliability-0-9.R
\name{process_measure_registry}
\alias{process_measure_registry}
\title{Unified process-measure registry}
\usage{
process_measure_registry(include_experimental = TRUE)
}
\arguments{
\item{include_experimental}{Include experimental registry entries.}
}
\description{
Unified process-measure registry
}
