% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/062-context-process-structure-0-8.R
\name{process_profile_probabilities}
\alias{process_profile_probabilities}
\title{Extract process-profile probabilities}
\usage{
process_profile_probabilities(x)
}
\arguments{
\item{x}{Object to process, inspect, compare, or plot.}
}
\description{
Extract process-profile probabilities
}
