% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/051-advanced-process-irt-0-7.R
\name{process_residual_map}
\alias{process_residual_map}
\title{Return person/item latent-space coordinates}
\usage{
process_residual_map(object, entity = c("both", "person", "item"))
}
\arguments{
\item{object}{A fitted eyeprocess model or audit object.}

\item{entity}{Entity type to map or validate.}
}
\description{
Return person/item latent-space coordinates
}
