% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/061-pupil-advanced-representations-0-8.R
\name{pupil_band_power}
\alias{pupil_band_power}
\title{Compute pupil signal power in a frequency band}
\usage{
pupil_band_power(y, sampling_rate_hz, lower_hz, upper_hz, detrend = TRUE)
}
\arguments{
\item{y}{Pupil signal.}

\item{sampling_rate_hz}{Sampling rate in Hz.}

\item{lower_hz, upper_hz}{Frequency-band limits.}

\item{detrend}{Remove the mean before FFT.}
}
\description{
Compute pupil signal power in a frequency band
}
