% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/061-pupil-advanced-representations-0-8.R
\name{pupil_event_effects}
\alias{pupil_event_effects}
\title{Extract event effects from pupil deconvolution}
\usage{
pupil_event_effects(x)
}
\arguments{
\item{x}{Object to process, inspect, compare, or plot.}
}
\description{
Extract event effects from pupil deconvolution
}
