% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/020-interoperability-storage.R
\name{sequence_interoperability}
\alias{sequence_interoperability}
\alias{as_procdata_sequence}
\alias{as_traminer_sequence}
\alias{as_seqhmm_data}
\title{Convert scanpaths to process/sequence package contracts}
\usage{
as_procdata_sequence(
  x,
  source = c("visits", "fixations", "samples"),
  collapse_consecutive = TRUE
)

as_traminer_sequence(
  x,
  source = c("visits", "fixations", "samples"),
  collapse_consecutive = TRUE,
  create_object = FALSE
)

as_seqhmm_data(
  x,
  source = c("visits", "fixations", "samples"),
  collapse_consecutive = TRUE
)
}
\arguments{
\item{x}{An `eye_dataset`.}

\item{source}{AOI sequence source.}

\item{collapse_consecutive}{Whether to collapse repeated adjacent states.}

\item{create_object}{For `as_traminer_sequence()`, whether to return a native
`TraMineR` sequence object instead of the package-neutral wide table.}
}
\value{
A package-compatible representation.
}
\description{
Convert scanpaths to process/sequence package contracts
}
