% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/063-operational-validation-0-8.R
\name{streaming_score_history}
\alias{streaming_score_history}
\title{Extract streaming score history}
\usage{
streaming_score_history(x)
}
\arguments{
\item{x}{Object to process, inspect, compare, or plot.}
}
\description{
Extract streaming score history
}
