% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/075-benchmark-stress-0-9.R
\name{summarise_eye_benchmark}
\alias{summarise_eye_benchmark}
\title{Summarise benchmark timing and memory by problem size}
\usage{
summarise_eye_benchmark(x)
}
\arguments{
\item{x}{Benchmark result.}
}
\value{
A logical value or vector indicating benchmark timing and memory by problem size.
}
\description{
Summarise benchmark timing and memory by problem size
}
