% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/060-process-window-representations-0-8.R
\name{summarize_process_windows}
\alias{summarize_process_windows}
\title{Summarize extracted process windows}
\usage{
summarize_process_windows(x, by = NULL)
}
\arguments{
\item{x}{`eye_process_windows` object.}

\item{by}{Optional grouping columns present in the extracted table.}
}
\description{
Summarize extracted process windows
}
