% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/078-reproducibility-provenance-0-9.R
\name{validate_eye_prov_graph}
\alias{validate_eye_prov_graph}
\title{Validate a provenance graph}
\usage{
validate_eye_prov_graph(x)
}
\arguments{
\item{x}{Provenance graph.}
}
\description{
Validate a provenance graph
}
