% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/087-irt-process-joint-0-9.R
\name{validate_eyeprocess_joint_process_irt_spec}
\alias{validate_eyeprocess_joint_process_irt_spec}
\title{Validate a joint process IRT specification}
\usage{
validate_eyeprocess_joint_process_irt_spec(x)
}
\arguments{
\item{x}{Object to validate, summarize, verify, or otherwise process.}
}
\description{
Validate a joint process IRT specification
}
