% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/019-gazepoint-downstream-workflow.R
\name{validate_gazepoint_workflow}
\alias{validate_gazepoint_workflow}
\title{Validate an integrated Gazepoint workflow result}
\usage{
validate_gazepoint_workflow(x)
}
\arguments{
\item{x}{An `eye_gazepoint_workflow` object.}
}
\value{
A data frame of workflow checks.
}
\description{
Validate an integrated Gazepoint workflow result
}
