% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/099-multimodal-validation-0-10.R
\name{validate_multimodal_irt}
\alias{validate_multimodal_irt}
\title{Validate a multimodal IRT development object}
\usage{
validate_multimodal_irt(x)
}
\arguments{
\item{x}{Measurement, fit, simulation, or process-information object.}
}
\value{
An `eye_multimodal_validation`.
}
\description{
Validate a multimodal IRT development object
}
