% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/069-validation-program-0-9.R
\name{validation_condition_id}
\alias{validation_condition_id}
\title{Return stable validation condition identifiers}
\usage{
validation_condition_id(x)
}
\arguments{
\item{x}{Validation design or expanded condition table.}
}
\description{
Return stable validation condition identifiers
}
