% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/069-validation-program-0-9.R
\name{validation_failure_profile}
\alias{validation_failure_profile}
\title{Failure profile for a validation programme}
\usage{
validation_failure_profile(x)
}
\arguments{
\item{x}{Validation result.}
}
\value{
A data frame containing failure profile for a validation programme. Rows represent the analysis units and columns contain the identifiers, estimates, or diagnostics defined by the function.
}
\description{
Failure profile for a validation programme
}
