% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/073-decision-manifests-0-9.R
\name{verify_decision_manifest_lock}
\alias{verify_decision_manifest_lock}
\title{Verify that a locked manifest has not changed}
\usage{
verify_decision_manifest_lock(x)
}
\arguments{
\item{x}{Manifest lock.}
}
\value{
A logical value or vector indicating verify that a locked manifest has not changed.
}
\description{
Verify that a locked manifest has not changed
}
