% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/078-reproducibility-provenance-0-9.R
\name{verify_reproducibility_fingerprint}
\alias{verify_reproducibility_fingerprint}
\title{Verify an internally stored fingerprint hash}
\usage{
verify_reproducibility_fingerprint(x)
}
\arguments{
\item{x}{Fingerprint.}
}
\value{
A logical value or vector indicating verify an internally stored fingerprint hash.
}
\description{
Verify an internally stored fingerprint hash
}
