% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/078-reproducibility-provenance-0-9.R
\name{write_prov_dot}
\alias{write_prov_dot}
\title{Return Graphviz DOT for a provenance graph}
\usage{
write_prov_dot(x)
}
\arguments{
\item{x}{Provenance graph.}
}
\description{
Return Graphviz DOT for a provenance graph
}
