% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/076-calibration-quality-uncertainty-0-9.R
\name{aoi_membership_probability}
\alias{aoi_membership_probability}
\title{AOI membership probabilities from uncertainty draws}
\usage{
aoi_membership_probability(draws, aois)
}
\arguments{
\item{draws}{Output of `propagate_calibration_uncertainty()` or compatible table.}

\item{aois}{Rectangular AOI table with aoi/x_min/x_max/y_min/y_max.}
}
\description{
AOI membership probabilities from uncertainty draws
}
