% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/071-api-lifecycle-0-9.R
\name{api_family_map}
\alias{api_family_map}
\title{Map exported APIs to conceptual families}
\usage{
api_family_map(inventory)
}
\arguments{
\item{inventory}{API inventory or character names.}
}
\description{
Map exported APIs to conceptual families
}
