% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/067-advanced-sensitivity-diagnostics-0-8.R
\name{compare_bayesian_process_models}
\alias{compare_bayesian_process_models}
\title{Compare Bayesian process models by LOO or Bayes factor}
\usage{
compare_bayesian_process_models(..., method = c("loo", "bayes_factor"))
}
\arguments{
\item{...}{Fitted brms models.}

\item{method}{`loo` or `bayes_factor`.}
}
\description{
Compare Bayesian process models by LOO or Bayes factor
}
