% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/076-calibration-quality-uncertainty-0-9.R
\name{compare_hard_probabilistic_aoi}
\alias{compare_hard_probabilistic_aoi}
\title{Compare hard and probabilistic AOI assignments}
\usage{
compare_hard_probabilistic_aoi(
  data,
  aois,
  probabilistic,
  x = "gaze_x",
  y = "gaze_y"
)
}
\arguments{
\item{data}{Gaze samples.}

\item{aois}{Rectangular AOIs.}

\item{probabilistic}{Probabilistic assignment object.}

\item{x, y}{Gaze-coordinate columns.}
}
\description{
Compare hard and probabilistic AOI assignments
}
