% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/062-context-process-structure-0-8.R
\name{compare_process_criterion_models}
\alias{compare_process_criterion_models}
\title{Compare process external-validity models}
\usage{
compare_process_criterion_models(x)
}
\arguments{
\item{x}{Object to process, inspect, compare, or plot.}
}
\description{
Compare process external-validity models
}
