% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/062-context-process-structure-0-8.R
\name{context_factor_effects}
\alias{context_factor_effects}
\title{Extract visual-context factor effects/loadings}
\usage{
context_factor_effects(x, IRTpars = FALSE)
}
\arguments{
\item{x}{Object to process, inspect, compare, or plot.}

\item{IRTpars}{Passed to the underlying IRT coefficient extractor to request IRT parameterization when supported.}
}
\description{
Extract visual-context factor effects/loadings
}
