% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/076-calibration-quality-uncertainty-0-9.R
\name{data_quality_reporting_table}
\alias{data_quality_reporting_table}
\title{Compact reporting table for eye-tracking data quality}
\usage{
data_quality_reporting_table(x)
}
\arguments{
\item{x}{Data-quality profile.}
}
\description{
Compact reporting table for eye-tracking data quality
}
