% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/073-decision-manifests-0-9.R
\name{decision_manifest_table}
\alias{decision_manifest_table}
\title{Flatten a decision manifest to a table}
\usage{
decision_manifest_table(x)
}
\arguments{
\item{x}{Manifest.}
}
\description{
Flatten a decision manifest to a table
}
