% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/070-governed-pipelines-0-9.R
\name{eye_analysis_pipeline}
\alias{eye_analysis_pipeline}
\title{Construct a governed analysis pipeline}
\usage{
eye_analysis_pipeline(
  ...,
  spec = eye_analysis_spec(),
  name = "eye_analysis",
  strict = TRUE
)
}
\arguments{
\item{...}{`eye_pipeline_step` objects.}

\item{spec}{Optional `eye_analysis_spec`.}

\item{name}{Pipeline label.}

\item{strict}{If `TRUE`, undeclared dependencies are errors.}
}
\description{
Construct a governed analysis pipeline
}
