% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/071-api-lifecycle-0-9.R
\name{eye_api_status}
\alias{eye_api_status}
\title{Lookup lifecycle status for one or more APIs}
\usage{
eye_api_status(name, registry = eye_api_lifecycle())
}
\arguments{
\item{name}{API names.}

\item{registry}{Lifecycle registry.}
}
\description{
Lookup lifecycle status for one or more APIs
}
