% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/071-api-lifecycle-0-9.R
\name{eye_api_superseded}
\alias{eye_api_superseded}
\title{Return superseded/deprecated compatibility interfaces}
\usage{
eye_api_superseded(registry = eye_api_lifecycle())
}
\arguments{
\item{registry}{Lifecycle registry.}
}
\description{
Return superseded/deprecated compatibility interfaces
}
