% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/070-governed-pipelines-0-9.R
\name{eye_pipeline_graph}
\alias{eye_pipeline_graph}
\title{Return pipeline vertices and dependency edges}
\usage{
eye_pipeline_graph(x)
}
\arguments{
\item{x}{Pipeline.}
}
\description{
Return pipeline vertices and dependency edges
}
