% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/070-governed-pipelines-0-9.R
\name{eye_pipeline_mermaid}
\alias{eye_pipeline_mermaid}
\title{Render pipeline dependencies as Mermaid flowchart text}
\usage{
eye_pipeline_mermaid(x)
}
\arguments{
\item{x}{Pipeline.}
}
\description{
Render pipeline dependencies as Mermaid flowchart text
}
