% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/090-irt-multidimensional-cdm-0-9.R
\name{eyeprocess_cdm_dina_probability}
\alias{eyeprocess_cdm_dina_probability}
\title{DINA response probabilities from slip and guess parameters}
\usage{
eyeprocess_cdm_dina_probability(ideal_response, slip = 0.1, guess = 0.2)
}
\arguments{
\item{ideal_response}{Ideal-response indicator or matrix implied by the cognitive-diagnosis model.}

\item{slip}{DINA slip parameter or vector of slip parameters.}

\item{guess}{DINA guessing parameter or vector of guessing parameters.}
}
\value{
An object of class "matrix", stored as an R object, containing dINA response probabilities from slip and guess parameters and associated metadata needed to interpret the result.
}
\description{
DINA response probabilities from slip and guess parameters
}
