% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/086-irt-linking-invariance-0-9.R
\name{eyeprocess_irt_apply_link}
\alias{eyeprocess_irt_apply_link}
\title{Apply linear IRT scale-linking coefficients}
\usage{
eyeprocess_irt_apply_link(items, link)
}
\arguments{
\item{items}{Item-parameter data frame or item collection.}

\item{link}{IRT scale-linking coefficients or linking object.}
}
\description{
Apply linear IRT scale-linking coefficients
}
