% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/083-irt-foundations-information-0-9.R
\name{eyeprocess_irt_conditional_sem}
\alias{eyeprocess_irt_conditional_sem}
\title{Conditional standard error from information}
\usage{
eyeprocess_irt_conditional_sem(information)
}
\arguments{
\item{information}{Item or test information value or vector.}
}
\value{
A logical value or vector indicating conditional standard error from information.
}
\description{
Conditional standard error from information
}
