% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/086-irt-linking-invariance-0-9.R
\name{eyeprocess_irt_dif_effect_curve}
\alias{eyeprocess_irt_dif_effect_curve}
\title{Differential item functioning effect curve from two parameter sets}
\usage{
eyeprocess_irt_dif_effect_curve(
  reference_item,
  focal_item,
  theta = seq(-4, 4, length.out = 81)
)
}
\arguments{
\item{reference_item}{Reference-group item parameters.}

\item{focal_item}{Focal-group item parameters.}

\item{theta}{Latent-trait value or vector of latent-trait values.}
}
\value{
An object of class "eye_irt_dif_curve", "data.frame", stored as a data frame, containing differential item functioning effect curve from two parameter sets and associated metadata needed to interpret the result.
}
\description{
Differential item functioning effect curve from two parameter sets
}
