% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/086-irt-linking-invariance-0-9.R
\name{eyeprocess_irt_dtf_curve}
\alias{eyeprocess_irt_dtf_curve}
\title{Differential test functioning effect curve}
\usage{
eyeprocess_irt_dtf_curve(reference, focal, theta = seq(-4, 4, length.out = 81))
}
\arguments{
\item{reference}{Reference-form or reference-group item parameters.}

\item{focal}{Focal-form or focal-group item parameters.}

\item{theta}{Latent-trait value or vector of latent-trait values.}
}
\description{
Differential test functioning effect curve
}
