% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/085-irt-scoring-adaptive-0-9.R
\name{eyeprocess_irt_eap_score}
\alias{eyeprocess_irt_eap_score}
\title{EAP score for dichotomous IRT item parameters}
\usage{
eyeprocess_irt_eap_score(
  response,
  items,
  theta_grid = seq(-4, 4, length.out = 81),
  prior_mean = 0,
  prior_sd = 1,
  D = 1
)
}
\arguments{
\item{response}{Observed item response or response variable.}

\item{items}{Item-parameter data frame or item collection.}

\item{theta_grid}{Grid of latent-trait values used for numerical scoring or integration.}

\item{prior_mean}{Mean of the normal latent-trait prior.}

\item{prior_sd}{Standard deviation of the normal latent-trait prior.}

\item{D}{Logistic scaling constant.}
}
\description{
EAP score for dichotomous IRT item parameters
}
