% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/088-irt-engine-adapters-0-9.R
\name{eyeprocess_irt_engine_status}
\alias{eyeprocess_irt_engine_status}
\title{Query an external IRT engine}
\usage{
eyeprocess_irt_engine_status(engine)
}
\arguments{
\item{engine}{Requested estimation or analysis engine.}
}
\description{
Query an external IRT engine
}
