% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/086-irt-linking-invariance-0-9.R
\name{eyeprocess_irt_functioning_effect_summary}
\alias{eyeprocess_irt_functioning_effect_summary}
\title{Summarise DIF/DTF curve magnitude}
\usage{
eyeprocess_irt_functioning_effect_summary(curve)
}
\arguments{
\item{curve}{Curve data to summarize or inspect.}
}
\description{
Summarise DIF/DTF curve magnitude
}
