% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/083-irt-foundations-information-0-9.R
\name{eyeprocess_irt_grm_probability}
\alias{eyeprocess_irt_grm_probability}
\title{Graded-response category probabilities}
\usage{
eyeprocess_irt_grm_probability(theta, a = 1, thresholds, D = 1)
}
\arguments{
\item{theta}{Latent-trait value or vector of latent-trait values.}

\item{a}{Item discrimination parameter or vector.}

\item{thresholds}{Ordered response-category thresholds.}

\item{D}{Logistic scaling constant.}
}
\description{
Graded-response category probabilities
}
