% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/083-irt-foundations-information-0-9.R
\name{eyeprocess_irt_identification_audit}
\alias{eyeprocess_irt_identification_audit}
\title{Audit IRT scale/location identification}
\usage{
eyeprocess_irt_identification_audit(
  spec,
  constraints = list(),
  n_items = NULL,
  n_persons = NULL
)
}
\arguments{
\item{spec}{Model, validation, or analysis specification object.}

\item{constraints}{Identification or model constraints.}

\item{n_items}{Number of items.}

\item{n_persons}{Number of persons.}
}
\description{
Audit IRT scale/location identification
}
