% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/083-irt-foundations-information-0-9.R
\name{eyeprocess_irt_information_area}
\alias{eyeprocess_irt_information_area}
\title{Area under an information curve}
\usage{
eyeprocess_irt_information_area(theta, information)
}
\arguments{
\item{theta}{Latent-trait value or vector of latent-trait values.}

\item{information}{Item or test information value or vector.}
}
\value{
A numeric value or vector containing area under an information curve.
}
\description{
Area under an information curve
}
