% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/085-irt-scoring-adaptive-0-9.R
\name{eyeprocess_irt_item_bank}
\alias{eyeprocess_irt_item_bank}
\title{Item bank object for adaptive design}
\usage{
eyeprocess_irt_item_bank(items, content = NULL, exposure_limit = 1)
}
\arguments{
\item{items}{Item-parameter data frame or item collection.}

\item{content}{Item content/category metadata.}

\item{exposure_limit}{Maximum permitted item exposure.}
}
\value{
An object of class "eye_irt_item_bank", stored as a named list, with components "items", "exposure_limit". It contains item bank object for adaptive design and associated metadata or diagnostics needed to interpret the result.
}
\description{
Item bank object for adaptive design
}
