% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/083-irt-foundations-information-0-9.R
\name{eyeprocess_irt_nominal_probability}
\alias{eyeprocess_irt_nominal_probability}
\title{Nominal-response category probabilities}
\usage{
eyeprocess_irt_nominal_probability(theta, slopes, intercepts)
}
\arguments{
\item{theta}{Latent-trait value or vector of latent-trait values.}

\item{slopes}{Nominal-category slope parameters.}

\item{intercepts}{Nominal-category intercept parameters.}
}
\value{
A numeric value or vector containing nominal-response category probabilities.
}
\description{
Nominal-response category probabilities
}
