% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/094-validation-paper-tables-atlas-0-9.R
\name{eyeprocess_irt_precision_evidence_table}
\alias{eyeprocess_irt_precision_evidence_table}
\title{Build a paper-ready IRT information/precision table}
\usage{
eyeprocess_irt_precision_evidence_table(
  items,
  theta = seq(-3, 3, by = 0.5),
  digits = 4L
)
}
\arguments{
\item{items}{Item-parameter data frame or item collection.}

\item{theta}{Latent-trait value or vector of latent-trait values.}

\item{digits}{Number of decimal digits used for presentation.}
}
\description{
Build a paper-ready IRT information/precision table
}
