% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/087-irt-process-joint-0-9.R
\name{eyeprocess_irt_process_alignment}
\alias{eyeprocess_irt_process_alignment}
\title{Align item parameters with process-channel summaries}
\usage{
eyeprocess_irt_process_alignment(
  item_parameters,
  process_profile,
  process_columns = NULL
)
}
\arguments{
\item{item_parameters}{Item-parameter data frame.}

\item{process_profile}{Process-measure profile or summary.}

\item{process_columns}{Names of process-measure columns to use.}
}
\description{
Align item parameters with process-channel summaries
}
