% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/085-irt-scoring-adaptive-0-9.R
\name{eyeprocess_irt_score_uncertainty}
\alias{eyeprocess_irt_score_uncertainty}
\title{Summarise score uncertainty}
\usage{
eyeprocess_irt_score_uncertainty(scores)
}
\arguments{
\item{scores}{Score object or score table.}
}
\description{
Summarise score uncertainty
}
