% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/083-irt-foundations-information-0-9.R
\name{eyeprocess_irt_test_characteristic_curve}
\alias{eyeprocess_irt_test_characteristic_curve}
\title{Test characteristic curve for dichotomous item parameters}
\usage{
eyeprocess_irt_test_characteristic_curve(theta, items, D = 1)
}
\arguments{
\item{theta}{Latent-trait value or vector of latent-trait values.}

\item{items}{Item-parameter data frame or item collection.}

\item{D}{Logistic scaling constant.}
}
\description{
Test characteristic curve for dichotomous item parameters
}
