% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/090-irt-multidimensional-cdm-0-9.R
\name{eyeprocess_irt_testlet_spec}
\alias{eyeprocess_irt_testlet_spec}
\title{Declare a testlet structure for bifactor/two-tier IRT engines}
\usage{
eyeprocess_irt_testlet_spec(item_id, testlet, general_dimension = "general")
}
\arguments{
\item{item_id}{Item identifier or vector of item identifiers.}

\item{testlet}{Testlet membership identifier.}

\item{general_dimension}{General dimension name or index.}
}
\value{
An object of class "eye_irt_testlet_spec", "data.frame", stored as a data frame, containing declare a testlet structure for bifactor/two-tier IRT engines and associated metadata needed to interpret the result.
}
\description{
Declare a testlet structure for bifactor/two-tier IRT engines
}
