% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/087-irt-process-joint-0-9.R
\name{eyeprocess_multichannel_measurement_map}
\alias{eyeprocess_multichannel_measurement_map}
\title{Construct a multichannel measurement map}
\usage{
eyeprocess_multichannel_measurement_map(
  response = "accuracy",
  channels = c("response_time", "dwell", "pupil", "transitions"),
  role = NULL
)
}
\arguments{
\item{response}{Observed item response or response variable.}

\item{channels}{Names or definitions of measurement channels.}

\item{role}{Declared role of each measurement channel.}
}
\description{
Construct a multichannel measurement map
}
