% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/087-irt-process-joint-0-9.R
\name{eyeprocess_process_person_profile}
\alias{eyeprocess_process_person_profile}
\title{Aggregate process channels by person}
\usage{
eyeprocess_process_person_profile(data, person, channels)
}
\arguments{
\item{data}{Input data frame, matrix, or compatible analysis object.}

\item{person}{Name of the person identifier column.}

\item{channels}{Names or definitions of measurement channels.}
}
\value{
A tabular R object containing aggregate process channels by person; rows represent analysis units and columns contain the returned quantities.
}
\description{
Aggregate process channels by person
}
