% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/087-irt-process-joint-0-9.R
\name{eyeprocess_response_time_profile}
\alias{eyeprocess_response_time_profile}
\title{Summarise response-time structure for joint IRT work}
\usage{
eyeprocess_response_time_profile(data, person, item, response_time)
}
\arguments{
\item{data}{Input data frame, matrix, or compatible analysis object.}

\item{person}{Name of the person identifier column.}

\item{item}{Name of the item identifier column.}

\item{response_time}{Response-time variable or values.}
}
\description{
Summarise response-time structure for joint IRT work
}
