% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/060-process-window-representations-0-8.R
\name{fit_aoi_growth_curve}
\alias{fit_aoi_growth_curve}
\title{Fit a single AOI growth curve}
\usage{
fit_aoi_growth_curve(data, time, outcome, degree = 3L)
}
\arguments{
\item{data}{Data frame.}

\item{time}{Time column.}

\item{outcome}{Numeric AOI proportion/indicator column.}

\item{degree}{Polynomial degree.}
}
\description{
Fit a single AOI growth curve
}
