% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/022-advanced-models-v2.R
\name{fit_gaze_diffusion_irt}
\alias{fit_gaze_diffusion_irt}
\title{Fit a gaze-informed diffusion/accumulation workflow}
\usage{
fit_gaze_diffusion_irt(x, spec = gaze_diffusion_spec(), ...)
}
\arguments{
\item{x}{An `eye_dataset`.}

\item{spec}{Diffusion specification.}

\item{...}{Passed to the selected engine.}
}
\value{
An `eye_gaze_diffusion_irt` object.
}
\description{
Fit a gaze-informed diffusion/accumulation workflow
}
