% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/022-advanced-models-v2.R
\name{fit_joint_functional_pupil_irt}
\alias{fit_joint_functional_pupil_irt}
\title{Fit a functional pupil-informed IRT workflow}
\usage{
fit_joint_functional_pupil_irt(x, spec = functional_pupil_irt_spec(), ...)
}
\arguments{
\item{x}{An `eye_dataset`.}

\item{spec}{Functional pupil-IRT specification.}

\item{...}{Passed to the selected model engine.}
}
\value{
An `eye_functional_pupil_irt` object.
}
\description{
Fit a functional pupil-informed IRT workflow
}
