% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/022-advanced-models-v2.R
\name{functional_pupil_irt_spec}
\alias{functional_pupil_irt_spec}
\title{Specify a functional pupil-IRT model}
\usage{
functional_pupil_irt_spec(
  df = 5L,
  response = "score",
  engine = c("two_stage_glm", "two_stage_lme4", "brms"),
  include_response_time = TRUE
)
}
\arguments{
\item{df}{Natural-spline degrees of freedom.}

\item{response}{Response field.}

\item{engine}{Two-stage GLM/multilevel engine or joint `brms` engine.}

\item{include_response_time}{Include a response-time submodel when joint.}
}
\value{
An `eye_functional_pupil_irt_spec`.
}
\description{
Specify a functional pupil-IRT model
}
