% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/076-calibration-quality-uncertainty-0-9.R
\name{gaze_precision_rms_s2s}
\alias{gaze_precision_rms_s2s}
\title{Estimate RMS successive-sample gaze imprecision}
\usage{
gaze_precision_rms_s2s(
  data,
  x = "gaze_x",
  y = "gaze_y",
  time = NULL,
  by = NULL
)
}
\arguments{
\item{data}{Gaze samples.}

\item{x, y}{Gaze-coordinate columns.}

\item{time}{Optional timestamp column used to order samples.}

\item{by}{Optional grouping columns.}
}
\description{
Estimate RMS successive-sample gaze imprecision
}
