% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/049-multimodal-irt-registry.R
\name{get_irt_model}
\alias{get_irt_model}
\title{Retrieve a registered multimodal IRT model}
\usage{
get_irt_model(id)
}
\arguments{
\item{id}{Stable identifier.}
}
\description{
Retrieve a registered multimodal IRT model
}
