% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/075-benchmark-stress-0-9.R
\name{inject_eye_missingness}
\alias{inject_eye_missingness}
\title{Inject generic missingness into selected columns}
\usage{
inject_eye_missingness(data, columns, proportion, seed = 1L)
}
\arguments{
\item{data}{Data.}

\item{columns}{Columns.}

\item{proportion}{Missingness proportion.}

\item{seed}{Seed.}
}
\description{
Inject generic missingness into selected columns
}
