% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/052-irt-validation-0-7.R
\name{leave_device_out_validation}
\alias{leave_device_out_validation}
\title{Leave device out validation}
\usage{
leave_device_out_validation(data, device, fitter, predictor, scorer)
}
\arguments{
\item{data}{Input data frame or compatible tabular object.}

\item{device}{Device identifier or device facet.}

\item{fitter}{Model-fitting function.}

\item{predictor}{Prediction function.}

\item{scorer}{Function that scores predictions.}
}
\description{
Leave device out validation
}
