% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/052-irt-validation-0-7.R
\name{leave_session_out_validation}
\alias{leave_session_out_validation}
\title{Leave session out validation}
\usage{
leave_session_out_validation(data, session, fitter, predictor, scorer)
}
\arguments{
\item{data}{Input data frame or compatible tabular object.}

\item{session}{Session identifier or session facet.}

\item{fitter}{Model-fitting function.}

\item{predictor}{Prediction function.}

\item{scorer}{Function that scores predictions.}
}
\value{
An object of class "eye_leave_session_out_validation", "data.frame", stored as a data frame, containing leave session out validation and associated metadata needed to interpret the result.
}
\description{
Leave session out validation
}
