% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/052-irt-validation-0-7.R
\name{leave_site_out_validation}
\alias{leave_site_out_validation}
\title{Leave site out validation}
\usage{
leave_site_out_validation(data, site, fitter, predictor, scorer)
}
\arguments{
\item{data}{Input data frame or compatible tabular object.}

\item{site}{Site identifier or site facet.}

\item{fitter}{Model-fitting function.}

\item{predictor}{Prediction function.}

\item{scorer}{Function that scores predictions.}
}
\description{
Leave site out validation
}
