% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/062-context-process-structure-0-8.R
\name{multiblock_person_coordinates}
\alias{multiblock_person_coordinates}
\title{Extract multiblock person coordinates}
\usage{
multiblock_person_coordinates(x)
}
\arguments{
\item{x}{Object to process, inspect, compare, or plot.}
}
\description{
Extract multiblock person coordinates
}
