% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/050-process-irt-models-0-7.R
\name{option_process_information}
\alias{option_process_information}
\title{Quantify option-process information from a nominal gaze model}
\usage{
option_process_information(object)
}
\arguments{
\item{object}{A fitted eyeprocess model or audit object.}
}
\value{
An object of class "eye_option_process_information", "data.frame", stored as a data frame, containing quantify option-process information from a nominal gaze model and associated metadata needed to interpret the result.
}
\description{
Quantify option-process information from a nominal gaze model
}
