% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/070-governed-pipelines-0-9.R
\name{pipeline_failures}
\alias{pipeline_failures}
\title{Return failed pipeline steps}
\usage{
pipeline_failures(x)
}
\arguments{
\item{x}{Pipeline run.}
}
\description{
Return failed pipeline steps
}
