% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/064-next-generation-plots-0-8.R
\name{plot.eye_biometric_preflight}
\alias{plot.eye_biometric_preflight}
\title{Plot biometric preflight diagnostics}
\usage{
\method{plot}{eye_biometric_preflight}(x, type = c("heatmap", "decision_counts"), ...)
}
\arguments{
\item{x}{Object to process, inspect, compare, or plot.}

\item{type}{Type of summary or visual representation to produce.}

\item{...}{Additional arguments passed to the underlying method or helper.}
}
\description{
Plot biometric preflight diagnostics
}
