% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/066-process-decision-features-0-8.R
\name{plot.eye_decision_process_proxy}
\alias{plot.eye_decision_process_proxy}
\title{Plot decision process proxy diagnostics}
\usage{
\method{plot}{eye_decision_process_proxy}(x, ...)
}
\arguments{
\item{x}{Object to process, inspect, compare, or plot.}

\item{...}{Additional arguments passed to the underlying method or helper.}
}
\description{
Plot decision process proxy diagnostics
}
