% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/064-next-generation-plots-0-8.R
\name{plot.eye_item_parameter_seed}
\alias{plot.eye_item_parameter_seed}
\title{Plot item parameter seed diagnostics}
\usage{
\method{plot}{eye_item_parameter_seed}(x, candidate_data = NULL, ...)
}
\arguments{
\item{x}{Object to process, inspect, compare, or plot.}

\item{candidate_data}{Optional candidate-item data used for the requested display.}

\item{...}{Additional arguments passed to the underlying method or helper.}
}
\description{
Plot item parameter seed diagnostics
}
