% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/053-process-irt-plots-0-7.R
\name{plot.eye_process_cat_simulation}
\alias{plot.eye_process_cat_simulation}
\title{Plot CAT simulation information accumulation}
\usage{
\method{plot}{eye_process_cat_simulation}(x, ...)
}
\arguments{
\item{x}{Object to print, plot, summarize, or audit.}

\item{...}{Additional arguments passed to the selected model, engine, or method.}
}
\description{
Plot CAT simulation information accumulation
}
