% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/064-next-generation-plots-0-8.R
\name{plot.eye_process_external_validity}
\alias{plot.eye_process_external_validity}
\title{Plot process external validity diagnostics}
\usage{
\method{plot}{eye_process_external_validity}(
  x,
  type = c("associations", "incremental", "observed_fitted", "residuals"),
  ...
)
}
\arguments{
\item{x}{Object to process, inspect, compare, or plot.}

\item{type}{Type of summary or visual representation to produce.}

\item{...}{Additional arguments passed to the underlying method or helper.}
}
\description{
Plot process external validity diagnostics
}
