% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/056-completion-plots-0-7.R
\name{plot.eye_process_facet_effects}
\alias{plot.eye_process_facet_effects}
\title{Plot eye process facet effects}
\usage{
\method{plot}{eye_process_facet_effects}(x, ...)
}
\arguments{
\item{x}{Object to print, plot, summarize, or audit.}

\item{...}{Additional arguments passed to the selected model, engine, or method.}
}
\description{
Plot eye process facet effects
}
