% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/053-process-irt-plots-0-7.R
\name{plot.eye_process_hmm_irt}
\alias{plot.eye_process_hmm_irt}
\title{Plot a process-HMM IRT fit}
\usage{
\method{plot}{eye_process_hmm_irt}(x, type = c("occupancy", "transition"), ...)
}
\arguments{
\item{x}{Object to print, plot, summarize, or audit.}

\item{type}{Plot or result type.}

\item{...}{Additional arguments passed to the selected model, engine, or method.}
}
\description{
Plot a process-HMM IRT fit
}
