% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/064-next-generation-plots-0-8.R
\name{plot.eye_process_window_sensitivity}
\alias{plot.eye_process_window_sensitivity}
\title{Plot process window sensitivity diagnostics}
\usage{
\method{plot}{eye_process_window_sensitivity}(x, ...)
}
\arguments{
\item{x}{Object to process, inspect, compare, or plot.}

\item{...}{Additional arguments passed to the underlying method or helper.}
}
\description{
Plot process window sensitivity diagnostics
}
