% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/064-next-generation-plots-0-8.R
\name{plot.eye_pupil_fatigue_drift}
\alias{plot.eye_pupil_fatigue_drift}
\title{Plot pupil fatigue drift diagnostics}
\usage{
\method{plot}{eye_pupil_fatigue_drift}(x, ...)
}
\arguments{
\item{x}{Object to process, inspect, compare, or plot.}

\item{...}{Additional arguments passed to the underlying method or helper.}
}
\description{
Plot pupil fatigue drift diagnostics
}
