% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/064-next-generation-plots-0-8.R
\name{plot.eye_signal_filter_audit}
\alias{plot.eye_signal_filter_audit}
\title{Plot signal filter audit diagnostics}
\usage{
\method{plot}{eye_signal_filter_audit}(x, ...)
}
\arguments{
\item{x}{Object to process, inspect, compare, or plot.}

\item{...}{Additional arguments passed to the underlying method or helper.}
}
\description{
Plot signal filter audit diagnostics
}
