% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/055-requested-api-completion-0-7.R
\name{plot_process_changepoint}
\alias{plot_process_changepoint}
\title{Plot detected process change points}
\usage{
plot_process_changepoint(object, ...)
}
\arguments{
\item{object}{An `eye_irt_changepoints`, `eye_changepoint_rt_irt`, or
`eye_changepoint_multimodal_irt` object.}

\item{...}{Graphical arguments.}
}
\description{
Plot detected process change points
}
