% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/064-next-generation-plots-0-8.R
\name{plot_pupil_activity_sensitivity}
\alias{plot_pupil_activity_sensitivity}
\title{Plot pupil activity sensitivity to window length}
\usage{
plot_pupil_activity_sensitivity(x, feature = "pupil_frequency_contrast", ...)
}
\arguments{
\item{x}{`eye_pupil_frequency_stability` object.}

\item{feature}{Feature name.}

\item{...}{Additional arguments passed to the underlying method or helper.}
}
\description{
Plot pupil activity sensitivity to window length
}
