% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/064-next-generation-plots-0-8.R
\name{plot_pupil_spectrum}
\alias{plot_pupil_spectrum}
\title{Plot a pupil-signal power spectrum}
\usage{
plot_pupil_spectrum(signal, sampling_rate_hz, max_hz = sampling_rate_hz/2, ...)
}
\arguments{
\item{signal}{Numeric pupil signal.}

\item{sampling_rate_hz}{Sampling rate.}

\item{max_hz}{Maximum frequency shown; defaults to Nyquist.}

\item{...}{Additional arguments passed to the underlying method or helper.}
}
\description{
Plot a pupil-signal power spectrum
}
