% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/060-process-window-representations-0-8.R
\name{predict_aoi_trajectory}
\alias{predict_aoi_trajectory}
\title{Predict from an AOI growth curve}
\usage{
predict_aoi_trajectory(object, time = NULL)
}
\arguments{
\item{object}{Object supplied to the S3 method.}

\item{time}{Time values or name of the time variable.}
}
\description{
Predict from an AOI growth curve
}
