% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/058-process-preflight-governance-0-8.R
\name{preflight_decisions}
\alias{preflight_decisions}
\title{Extract pre-flight decisions}
\usage{
preflight_decisions(x)
}
\arguments{
\item{x}{An `eye_biometric_preflight` object.}
}
\description{
Extract pre-flight decisions
}
