% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/058-process-preflight-governance-0-8.R
\name{process_anomaly_distance}
\alias{process_anomaly_distance}
\title{Extract multivariate process anomaly distances}
\usage{
process_anomaly_distance(x)
}
\arguments{
\item{x}{Object to process, inspect, compare, or plot.}
}
\description{
Extract multivariate process anomaly distances
}
