% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/051-advanced-process-irt-0-7.R
\name{process_dif_nuisance_surrogate}
\alias{process_dif_nuisance_surrogate}
\title{Construct a process-data nuisance surrogate for DIF analysis}
\usage{
process_dif_nuisance_surrogate(
  data,
  process_features,
  person = "participant_id",
  aggregate = TRUE
)
}
\arguments{
\item{data}{Input data frame or compatible tabular object.}

\item{process_features}{Names of process-derived features.}

\item{person}{Person or participant identifier column.}

\item{aggregate}{Aggregation rule for process features.}
}
\description{
Construct a process-data nuisance surrogate for DIF analysis
}
