% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/074-process-registry-reliability-0-9.R
\name{process_icc}
\alias{process_icc}
\title{Absolute-agreement ICC(A,1) for repeated process measures}
\usage{
process_icc(data, person, session, measure)
}
\arguments{
\item{data}{Long data.}

\item{person}{Participant column.}

\item{session}{Session/repetition column.}

\item{measure}{Measure column.}
}
\description{
Absolute-agreement ICC(A,1) for repeated process measures
}
