% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/074-process-registry-reliability-0-9.R
\name{process_measure_card}
\alias{process_measure_card}
\title{Return a one-measure process card}
\usage{
process_measure_card(name, registry = process_measure_registry())
}
\arguments{
\item{name}{Measure name.}

\item{registry}{Registry.}
}
\description{
Return a one-measure process card
}
