% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/074-process-registry-reliability-0-9.R
\name{process_measure_coverage}
\alias{process_measure_coverage}
\title{Process-measure coverage for an observed dataset}
\usage{
process_measure_coverage(data, registry = process_measure_registry())
}
\arguments{
\item{data}{Data frame.}

\item{registry}{Registry.}
}
\description{
Process-measure coverage for an observed dataset
}
