% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/074-process-registry-reliability-0-9.R
\name{process_measure_guardrails}
\alias{process_measure_guardrails}
\title{Process-measure guardrail table}
\usage{
process_measure_guardrails(registry = process_measure_registry())
}
\arguments{
\item{registry}{Registry.}
}
\description{
Process-measure guardrail table
}
