% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/062-context-process-structure-0-8.R
\name{process_profile_summary}
\alias{process_profile_summary}
\title{Summarize process profiles}
\usage{
process_profile_summary(x)
}
\arguments{
\item{x}{Object to process, inspect, compare, or plot.}
}
\description{
Summarize process profiles
}
