% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/074-process-registry-reliability-0-9.R
\name{process_reliability_profile}
\alias{process_reliability_profile}
\title{Test-retest process reliability profile}
\usage{
process_reliability_profile(data, person, session, measure)
}
\arguments{
\item{data}{Long data.}

\item{person, session, measure}{Column names.}
}
\description{
Test-retest process reliability profile
}
