% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/051-advanced-process-irt-0-7.R
\name{process_state_occupancy}
\alias{process_state_occupancy}
\title{Summarize HMM state occupancy}
\usage{
process_state_occupancy(object)
}
\arguments{
\item{object}{A fitted eyeprocess model or audit object.}
}
\description{
Summarize HMM state occupancy
}
