% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/051-advanced-process-irt-0-7.R
\name{process_state_transition_summary}
\alias{process_state_transition_summary}
\title{Summarize HMM process-state transitions}
\usage{
process_state_transition_summary(object)
}
\arguments{
\item{object}{A fitted eyeprocess model or audit object.}
}
\description{
Summarize HMM process-state transitions
}
