% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/074-process-registry-reliability-0-9.R
\name{process_temporal_stability}
\alias{process_temporal_stability}
\title{Pairwise temporal stability across sessions}
\usage{
process_temporal_stability(
  data,
  person,
  session,
  measure,
  method = c("pearson", "spearman")
)
}
\arguments{
\item{data}{Long data.}

\item{person, session, measure}{Column names.}

\item{method}{Correlation method.}
}
\value{
A logical value or vector indicating pairwise temporal stability across sessions.
}
\description{
Pairwise temporal stability across sessions
}
