% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/061-pupil-advanced-representations-0-8.R
\name{pupil_confound_effects}
\alias{pupil_confound_effects}
\title{Extract pupil confound-model effects}
\usage{
pupil_confound_effects(x)
}
\arguments{
\item{x}{Object to process, inspect, compare, or plot.}
}
\description{
Extract pupil confound-model effects
}
