% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/061-pupil-advanced-representations-0-8.R
\name{pupil_event_regressor}
\alias{pupil_event_regressor}
\title{Build an event-locked pupil regressor}
\usage{
pupil_event_regressor(time_ms, event_time_ms, tmax_ms = 930, shape = 10.1)
}
\arguments{
\item{time_ms}{Sample times.}

\item{event_time_ms}{Event onset.}

\item{tmax_ms, shape}{Kernel parameters.}
}
\description{
Build an event-locked pupil regressor
}
