% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/061-pupil-advanced-representations-0-8.R
\name{pupil_response_kernel}
\alias{pupil_response_kernel}
\title{Canonical gamma-shaped pupil response kernel}
\usage{
pupil_response_kernel(
  time_since_event_ms,
  tmax_ms = 930,
  shape = 10.1,
  normalize = TRUE
)
}
\arguments{
\item{time_since_event_ms}{Time relative to event onset.}

\item{tmax_ms}{Approximate response peak time.}

\item{shape}{Shape parameter.}

\item{normalize}{Normalize peak to one.}
}
\description{
Canonical gamma-shaped pupil response kernel
}
