% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/061-pupil-advanced-representations-0-8.R
\name{pupil_velocity_activity}
\alias{pupil_velocity_activity}
\title{Derivative-based pupil activity magnitude}
\usage{
pupil_velocity_activity(y, time_ms)
}
\arguments{
\item{y}{Pupil signal.}

\item{time_ms}{Time in milliseconds.}
}
\description{
Derivative-based pupil activity magnitude
}
