% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/070-governed-pipelines-0-9.R
\name{resume_eye_pipeline}
\alias{resume_eye_pipeline}
\title{Resume a governed pipeline from a prior run}
\usage{
resume_eye_pipeline(x, previous, context = list(), stop_on_error = TRUE)
}
\arguments{
\item{x}{Pipeline.}

\item{previous}{Prior pipeline run.}

\item{context}{Context used for new steps.}

\item{stop_on_error}{Stop on non-optional error.}
}
\description{
Resume a governed pipeline from a prior run
}
