% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/079-software-paper-evidence-0-9.R
\name{software_paper_gap_analysis}
\alias{software_paper_gap_analysis}
\title{Identify gaps in a software-paper evidence bundle}
\usage{
software_paper_gap_analysis(x)
}
\arguments{
\item{x}{Evidence bundle.}
}
\description{
Identify gaps in a software-paper evidence bundle
}
