% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/072-sensitivity-multiverse-0-9.R
\name{specification_coverage}
\alias{specification_coverage}
\title{Fraction of planned specifications successfully evaluated}
\usage{
specification_coverage(x)
}
\arguments{
\item{x}{Sensitivity result.}
}
\description{
Fraction of planned specifications successfully evaluated
}
