% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/072-sensitivity-multiverse-0-9.R
\name{specification_curve_data}
\alias{specification_curve_data}
\title{Prepare ordered specification-curve data}
\usage{
specification_curve_data(x, effect = "effect", lower = NULL, upper = NULL)
}
\arguments{
\item{x}{Sensitivity result.}

\item{effect}{Effect column.}

\item{lower}{Optional lower interval column.}

\item{upper}{Optional upper interval column.}
}
\description{
Prepare ordered specification-curve data
}
