% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/069-validation-program-0-9.R
\name{summarise_process_validation}
\alias{summarise_process_validation}
\title{Summarise a process-validation result}
\usage{
summarise_process_validation(x, by = NULL)
}
\arguments{
\item{x}{Validation result.}

\item{by}{Optional grouping variables in addition to parameter.}
}
\description{
Summarise a process-validation result
}
