% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/082-validation-evidence-programs-0-9.R
\name{summarise_validation_acceptance}
\alias{summarise_validation_acceptance}
\title{Summarise an acceptance matrix}
\usage{
summarise_validation_acceptance(x, by = character())
}
\arguments{
\item{x}{Object to validate, summarize, verify, or otherwise process.}

\item{by}{Grouping variables or aggregation level.}
}
\description{
Summarise an acceptance matrix
}
