% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/022-advanced-models-v2.R
\name{theory_strategy_spec}
\alias{theory_strategy_spec}
\title{Specify theory-defined process strategies}
\usage{
theory_strategy_spec(prototypes, feature_sd = NULL, prior = NULL)
}
\arguments{
\item{prototypes}{Matrix/data frame with strategies in rows and features in columns.}

\item{feature_sd}{Optional feature standard deviations used in distance likelihoods.}

\item{prior}{Optional strategy prior probabilities.}
}
\value{
An `eye_theory_strategy_spec`.
}
\description{
Specify theory-defined process strategies
}
