% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/070-governed-pipelines-0-9.R
\name{validate_eye_pipeline}
\alias{validate_eye_pipeline}
\title{Validate a governed eyeprocess pipeline}
\usage{
validate_eye_pipeline(x)
}
\arguments{
\item{x}{Pipeline.}
}
\description{
Validate a governed eyeprocess pipeline
}
