% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/083-irt-foundations-information-0-9.R
\name{validate_eyeprocess_irt_model_spec}
\alias{validate_eyeprocess_irt_model_spec}
\title{Validate an eyeprocess IRT model specification}
\usage{
validate_eyeprocess_irt_model_spec(x)
}
\arguments{
\item{x}{Object to validate, summarize, verify, or otherwise process.}
}
\value{
A logical value or vector indicating an eyeprocess IRT model specification.
}
\description{
Validate an eyeprocess IRT model specification
}
