% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/082-validation-evidence-programs-0-9.R
\name{validate_eyeprocess_validation_plan}
\alias{validate_eyeprocess_validation_plan}
\title{Validate a validation-evidence plan}
\usage{
validate_eyeprocess_validation_plan(x)
}
\arguments{
\item{x}{Object to validate, summarize, verify, or otherwise process.}
}
\value{
A logical value or vector indicating a validation-evidence plan.
}
\description{
Validate a validation-evidence plan
}
