% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/074-process-registry-reliability-0-9.R
\name{validate_process_measure_registry}
\alias{validate_process_measure_registry}
\title{Validate a process-measure registry}
\usage{
validate_process_measure_registry(registry)
}
\arguments{
\item{registry}{Registry data frame.}
}
\description{
Validate a process-measure registry
}
