% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/069-validation-program-0-9.R
\name{validate_process_validation_design}
\alias{validate_process_validation_design}
\title{Validate a process-validation design}
\usage{
validate_process_validation_design(x)
}
\arguments{
\item{x}{Validation design.}
}
\value{
A logical value or vector indicating a process-validation design.
}
\description{
Validate a process-validation design
}
