% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/069-validation-program-0-9.R
\name{validation_robustness_score}
\alias{validation_robustness_score}
\title{Overall validation robustness score}
\usage{
validation_robustness_score(x)
}
\arguments{
\item{x}{Validation result.}
}
\description{
Overall validation robustness score
}
