% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/073-decision-manifests-0-9.R
\name{verify_outcome_blind_snapshot}
\alias{verify_outcome_blind_snapshot}
\title{Verify an outcome-blind snapshot has not changed}
\usage{
verify_outcome_blind_snapshot(x)
}
\arguments{
\item{x}{Snapshot.}
}
\value{
A logical value or vector indicating verify an outcome-blind snapshot has not changed.
}
\description{
Verify an outcome-blind snapshot has not changed
}
